@@ -1,5 +1,5 @@
 module embeddoc
 
-go 1.25.6
+go 1.26.0
 
 require gopkg.in/yaml.v3 v3.0.1
