@@ -5,7 +5,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/utils/ptr"
 
 	"github.com/argoproj/argo-workflows/v4/pkg/apiclient/workflowarchive"
 )
@@ -23,7 +22,7 @@ func Test_flatten(t *testing.T) {
 			Watch:               false,
 			AllowWatchBookmarks: true,
 			ResourceVersion:     "11",
-			TimeoutSeconds:      ptr.To(int64(22)),
+			TimeoutSeconds:      new(int64(22)),
 			Limit:               33,
 			Continue:            "44",
 		}}, map[string]string{
