@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"k8s.io/utils/ptr"
 
 	"github.com/argoproj/argo-workflows/v4/pkg/apis/workflow/v1alpha1"
 	"github.com/argoproj/argo-workflows/v4/util/logging"
@@ -22,7 +21,7 @@ func TestKillDaemonChildrenUnmarkPod(t *testing.T) {
 				"a": v1alpha1.NodeStatus{
 					ID:         "a",
 					BoundaryID: "a",
-					Daemoned:   ptr.To(true),
+					Daemoned:   new(true),
 				},
 			},
 		},
