@@ -12,7 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/utils/ptr"
 )
 
 type MockKC struct {
@@ -59,7 +58,7 @@ func TestTerminatePodWithContainerName(t *testing.T) {
 				Name: "foo",
 			},
 			Spec: v1.PodSpec{
-				ShareProcessNamespace: ptr.To(true),
+				ShareProcessNamespace: new(true),
 			},
 		},
 		getContainerStatusContainerStatus: &v1.ContainerStatus{
