@@ -9,7 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
 	"k8s.io/client-go/util/workqueue"
-	"k8s.io/utils/ptr"
 
 	wfv1 "github.com/argoproj/argo-workflows/v4/pkg/apis/workflow/v1alpha1"
 	"github.com/argoproj/argo-workflows/v4/util/logging"
@@ -118,7 +117,7 @@ func TestRealtimeMetricGC(t *testing.T) {
 		Labels: labels,
 		Help:   "None",
 		Gauge: &wfv1.Gauge{
-			Realtime: ptr.To(true),
+			Realtime: new(true),
 		}},
 		wfKey,
 		func() float64 { return 1.0 },
@@ -196,7 +195,7 @@ func TestRealTimeMetricDeletion(t *testing.T) {
 		Help:   "hello",
 		Gauge: &wfv1.Gauge{
 			Value:     "1.0",
-			Realtime:  ptr.To(true),
+			Realtime:  new(true),
 			Operation: wfv1.GaugeOperationAdd,
 		},
 	}, "123", func() float64 { return 0.0 })
@@ -223,7 +222,7 @@ func TestRealTimeMetricDeletion(t *testing.T) {
 		Help:   "hello",
 		Gauge: &wfv1.Gauge{
 			Value:     "1.0",
-			Realtime:  ptr.To(true),
+			Realtime:  new(true),
 			Operation: wfv1.GaugeOperationAdd,
 		},
 	}, "456", nil)
